refactor: rename utilityType.ts -> apperance.ts
